requirements.txt
"""pip install -r requirements.txt"""

matplotlib
scikit-learn
numpy
faiss-cpu